Use binary, avoid patching source, etc
This reworks a lot, but primarily focuses on using a stable build instead of compiling source (saves CI/CD time). In lieu of this change, the base image was switched to the Home Assistant base image for debian instead of the addon debian base for simplicity in tagging (and a more recent image), and also some changes to how Hyperion is patched to support Ingress.